refactor: restructure user interface layer and fix cross-module issues
- Refactor user interface layer: split monolithic serializers.py and views.py into modular structure
  - Move user serializers to serializers/user_serializers.py
  - Move user views to views/user_views.py
  - Add new professional_views.py for professional-specific endpoints
  - Update __init__.py files to maintain backward compatibility

- Fix branding admin: correct BrandTheme list_display fields to match model

- Enhance catalog module:
  - Update PrestationViewModel to include account_id
  - Fix ListPrestations use case to pass account context
  - Update PrestationSerializer and views for account-aware filtering

- Improve user module:
  - Add new application errors (ProfessionalNotFoundError, ProfessionalAlreadyExistsError)
  - Add IsAccountOwner permission class for account-level access control

- Fix client tests: update test_create_client to use correct account fixture

- Update settings: add CORS and session configuration for Docker environment
