+++ backend/apps/branding/admin.py
@@ -6,13 +6,13 @@ from .models import BrandTheme
 
 @admin.register(BrandTheme)
 class BrandThemeAdmin(admin.ModelAdmin):
-    list_display = ("name", "professional", "is_active", "created_at", "updated_at")
-    search_fields = ("name", "professional__email")
+    list_display = ("name", "account", "is_active", "created_at", "updated_at")
+    search_fields = ("name", "account__email")
     list_filter = ("is_active",)
     ordering = ("-created_at",)
 
     fieldsets = (
-        ("Basic Information", {"fields": ("id", "professional", "name", "is_active")}),
+        ("Basic Information", {"fields": ("id", "account", "name", "is_active")}),
         ("Customization", {"fields": ("colors", "typography", "spacing", "logo")}),
         ("Timestamps", {"fields": ("created_at", "updated_at")}),
     )
