+++ backend/apps/catalog/application/dto/prestation_viewmodels.py
@@ -32,7 +32,7 @@ class PrestationViewModel:
     default_rate_display: str  # Format "1200.00"
     status: str
     custom: bool
-    professional_user_id: Optional[int]
+    account_id: Optional[int]  # Phase 5.4: renamed from professional_user_id
     created_at: datetime
     updated_at: datetime
