+++ backend/apps/catalog/application/usecases/list_prestations.py
@@ -67,7 +67,7 @@ class ListPrestations:
             default_rate_display=PrestationCalculator.format_rate_display(prestation.default_rate_cents),
             status=prestation.status,
             custom=prestation.custom,
-            professional_user_id=prestation.professional_user_id,
+            account_id=prestation.account_id,  # Phase 5.4
             created_at=prestation.created_at,
             updated_at=prestation.updated_at,
         )
