+++ backend/apps/client/tests/application/test_create_client.py
@@ -11,7 +11,8 @@ class InMemoryRepo:
     def create(self, data):
         obj = SimpleNamespace(
             id="uuid-1",
-            owner_id=data["owner_id"],
+            owner_id=data.get("owner_id"),
+            account_id=data.get("account_id"),
             name=data["name"],
             email=data["email"],
             phone=data["phone"],
@@ -24,12 +25,12 @@ class InMemoryRepo:
         self.saved.append(obj)
         return obj
 
-    def exists_by_owner_name(self, owner_id: int, name: str) -> bool:
-        return any(obj.owner_id == owner_id and obj.name == name for obj in self.saved)
+    def exists_by_account_name(self, account_id: int, name: str) -> bool:
+        return any(obj.account_id == account_id and obj.name == name for obj in self.saved)
 
 
 def test_create_client_minimal():
     uc = CreateClient(InMemoryRepo())
-    vm = uc.execute(CreateClientInput(owner_id=1, name=" ACME "))
+    vm = uc.execute(CreateClientInput(owner_id=1, account_id=1, name=" ACME "))
     assert vm.name == "ACME"
     assert vm.owner_id == 1
