+++ backend/apps/user/interface/serializers/__init__.py
@@ -1,7 +1,33 @@
-# apps/user/interface/serializers/__init__.py
 from apps.user.interface.serializers.account_serializers import (
     AccountInputSerializer,
     AccountOutputSerializer,
 )
+from apps.user.interface.serializers.user_serializers import (
+    ChangePasswordInputSerializer,
+    LogoutSerializer,
+    ProfessionalOutputSerializer,
+    ProfessionalUpsertInputSerializer,
+    ProfileOutputSerializer,
+    ProfilePatchInputSerializer,
+    RequestPasswordResetSerializer,
+    ResetPasswordSerializer,
+    UserListOutputSerializer,
+    UserOutputSerializer,
+    UserRegistrationInputSerializer,
+)
 
-__all__ = ["AccountInputSerializer", "AccountOutputSerializer"]
+__all__ = [
+    "AccountInputSerializer",
+    "AccountOutputSerializer",
+    "ChangePasswordInputSerializer",
+    "LogoutSerializer",
+    "ProfessionalOutputSerializer",
+    "ProfessionalUpsertInputSerializer",
+    "ProfileOutputSerializer",
+    "ProfilePatchInputSerializer",
+    "RequestPasswordResetSerializer",
+    "ResetPasswordSerializer",
+    "UserListOutputSerializer",
+    "UserOutputSerializer",
+    "UserRegistrationInputSerializer",
+]
