+++ backend/apps/user/interface/views/__init__.py
@@ -1,4 +1,13 @@
-# apps/user/interface/views/__init__.py
 from apps.user.interface.views.account_views import AccountViewSet
+from apps.user.interface.views.professional_views import (
+    OnboardingProfessionalView,
+    ProfessionalUserMeView,
+)
+from apps.user.interface.views.user_views import UserViewSet
 
-__all__ = ["AccountViewSet"]
+__all__ = [
+    "AccountViewSet",
+    "UserViewSet",
+    "ProfessionalUserMeView",
+    "OnboardingProfessionalView",
+]
