+++ docs/knowledge.md
@@ -1585,5 +1585,5 @@ _No package.json found at /Users/bertrandrenaudin/Desktop/DEV/FreelanSign/backen
 
 _Last updated_
 <!-- BEGIN AUTO: LAST_UPDATED -->
-_Updated_: **2025-11-30 14:17:36 CET**
+_Updated_: **2025-11-30 16:32:29 CET**
 <!-- END AUTO: LAST_UPDATED -->
